+++ sdk/dotnet/AlertChannel.cs
@@ -36,7 +36,7 @@ namespace Pulumi.Newrelic
         /// <param name="args">The arguments used to populate this resource's properties</param>
         /// <param name="options">A bag of options that control this resource's behavior</param>
         public AlertChannel(string name, AlertChannelArgs args, CustomResourceOptions? options = null)
-            : base("newrelic:index/alertChannel:AlertChannel", name, args, MakeResourceOptions(options, ""))
+            : base("newrelic:index/alertChannel:AlertChannel", name, args ?? ResourceArgs.Empty, MakeResourceOptions(options, ""))
         {
         }