+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Upgrade to support go 1.13.x
 * Upgrade to v1.8.0 of NewRelic Terraform Provider
+* Upgrade to pulumi-terraform@dotnet
 
 ---